refactor(cockpit): extract _FirstShippedCelebrationModal into sibling module (#1354)
Second wedge of the cockpit_ui.py god-module split (after #1606).

Moves the ~70-LOC _FirstShippedCelebrationModal plus its companion
_FIRST_SHIPPED_FRAMES tuple and _celebrate_first_shipped helper into a
new sibling module src/pollypm/cockpit_first_shipped.py. The whole
celebration surface is self-contained, so co-locating the helper keeps
the boundary clean rather than leaving the helper orphaned in
cockpit_ui.

Leaves an import shim in cockpit_ui.py so existing call sites and the
two-test suite tests/test_cockpit_first_shipped.py keep importing from
cockpit_ui unchanged.

cockpit_ui.py: 18945 -> 18870 LOC (-75); diff in cockpit_ui.py is 85
lines (well under the 200-line wedge budget).
